requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.1.0
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
tensorboardX==2.0
torch==1.4.0
torchvision==0.5.0
